added craiglist scraper to route
