+++ client/src/utils/API.js
@@ -74,6 +74,7 @@ export default {
     return axios.get("/api/walmart/" + searchTerm);
   },
   scrapeCraiglist: function (searchTerm) {
+    
     return axios.get("/api/Craiglist/" + searchTerm);
   }
