+++ routes/api/scraperCraig.js
@@ -1,0 +1,8 @@
+const router = require("express").Router();
+const scraperController = require("../../controllers/scrapeCraigController");
+
+router
+    .route("/Craiglist/:searchTerm")
+    .get(scrapeCraigController.scrapeCraiglist);
+
+module.exports = router;
